Add SecurityBaseConfig , upgrate User Mappers,

+++ src/main/java/sdainalproject/meetup/repository/UserRepository.java
@@ -1,8 +1,7 @@
 package sdainalproject.meetup.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
 import sdainalproject.meetup.dao.UserModel;
 
-public interface UserRepository extends JpaRepository<UserModel, Integer> {
+public interface UserRepository extends JpaRepository<UserModel, Long> {
 }
